+++ app/config.py
@@ -17,6 +17,7 @@ def load_config(
     connection_pool_name: str = "wwdtm_reports",
     app_time_zone: str = "UTC",
 ) -> dict[str, dict[str, Any]]:
+    """Load and parse configuration JSON."""
     _config_file = Path(config_file_path)
     with _config_file.open(mode="r", encoding="utf-8") as config_file:
         app_config = json.load(config_file)